Fix crashing/not working edition in VoxelLodTerrain

+++ generators/voxel_generator.h
@@ -38,7 +38,7 @@ public:
 	struct VoxelQueryData {
 		VoxelBufferInternal &voxel_buffer;
 		Vector3i origin_in_voxels;
-		uint8_t lod;
+		uint32_t lod;
 	};
 
 	virtual Result generate_block(VoxelQueryData &input);
